Rename modules to match naming conventions
